backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marclosk <marclosquino2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
django = "^5.1.2"
pytest = "^7.0"
pytest-django = "^4.0"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
asgiref = "^3.3.1"
psycopg2 = "^2.8.6"
psycopg2-binary = "^2.8.6"
pytz = "^2020.4"
sqlparse = "^0.4.1"
pillow = "^11.0.0"
gunicorn = "^20.1.0"
boto3 = "^1.37.3"
django-storages = "^1.14.5"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
